refactor: оптимизация сборки
