+++ package.json
@@ -34,7 +34,7 @@
         "lodash.clonedeep": "4.5.0",
         "long": "5.2.3",
         "protobufjs": "7.2.3",
-        "uuid": "8.0.0"
+        "uuid": "9.0.0"
     },
     "browserslist": [
         "last 1 Chrome versions"
