+++ src/record/version.ts
@@ -1,0 +1,1 @@
+export const CURRENT_VERSION = '0.1.0';
